+++ changelog.md
@@ -4,3 +4,7 @@
 
 #### 2024.05.12 (**0.1.0**):
 * Initial release.
+
+#### 2024.05.17 (**0.1.2**):
+* Fixes: AddFCProperties - selection.
+* Fixes: Explode - animation.